Move some utility functions intp pyjaia.utils (used by task_packet_database)
